Use strtoul instead of strtol for sid parsing. Fixes parsing of really large sid numbers. Fixes #393.

+++ src/detect-rev.c
@@ -61,7 +61,7 @@ static int DetectRevSetup (DetectEngineCtx *de_ctx, Signature *s, char *rawstr)
                    "to rev keyword");
         goto error;
     }
-    s->rev = rev;
+    s->rev = (uint32_t)rev;
 
     if (dubbed)
         SCFree(str);
